Update README.md
Introduction: The Graph Coloring Problem is a well-known combinatorial optimization challenge that involves
assigning colors to the vertices of a graph in such a way that no two adjacent vertices share the same color. This
project aims to develop an intelligent Graph Coloring Problem Solver utilizing both the Backtracking Search
Algorithm and a Genetic Algorithm.
Problem Description: In the Graph Coloring Problem, we have a graph composed of vertices and edges. The
objective is to color each vertex in a way that no two connected vertices have the same color. The minimum
number of colors required to achieve this is known as the chromatic number of the graph. Our task is to find a
valid colouring solution that minimises the chromatic number.
Backtracking Search Algorithm: The Backtracking Search Algorithm is a systematic exploration approach that
examines different colour assignments for each vertex. It works by recursively trying out colours and
backtracking when conflicts arise. This algorithm explores the solution space while intelligently pruning paths
that lead to infeasible colourings.
Genetic Algorithm: The Genetic Algorithm is an optimisation technique inspired by natural selection. In this
context, we represent potential colourings as chromosomes. The algorithm evolves a population of colourings
over generations using genetic operations such as crossover and mutation. The fittest colourings, those with
fewer conflicts, are more likely to be passed to the next generation.
USING GUI